# docs/api/openapi/sky-offer.openapi.yaml
openapi: "3.1.0"

info:
  title: sky-offer
  version: v1
  description: |
    Offer management service for the Sky platform. Provides public read access to offers
    and authenticated write operations for offer owners.

    Public endpoints (no token required):
    - `GET /offers` and `GET /offers/**` — browse all offers
    - `POST /search` — full-text search across offers

    All other endpoints require a valid Keycloak bearer JWT. The `email` claim identifies
    the owner; it is never sent in the request body for write operations.

    The `GET /offers/{offerId}/owner` endpoint is a service-to-service path consumed by
    sky-booking to resolve the offer owner when creating a booking. It is authenticated
    and reachable through the standard gateway route.

servers:
  - url: http://localhost:5777/offer/api
    description: Local via Spring Cloud Gateway (sky-gateway on port 5777)
  - url: https://skycloud.luksarna.com/offer/api
    description: Production (GCP via nginx-ingress)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Keycloak-issued JWT. Obtain via the password grant or client-credentials grant against
        `https://keycloak.luksarna.com/realms/sky/protocol/openid-connect/token`.
        The `email` claim in the token identifies the calling user as the offer owner for
        all write operations. The token is validated against the Keycloak JWKS endpoint.

  schemas:
    OfferDTO:
      type: object
      description: >
        Full offer representation used for create (`POST /owner/offers`) and all read
        responses. On create the `ownerEmail` field is ignored in the request body —
        the service sets it from the JWT `email` claim.
      required:
        - hotelName
        - price
        - roomCapacity
        - city
        - country
      properties:
        id:
          type: integer
          format: int64
          description: Auto-generated offer ID. Omit when creating; present in responses.
          example: 42
        hotelName:
          type: string
          description: Name of the hotel or accommodation. Must not be blank.
          example: "Grand Hotel Paris"
        description:
          type: string
          maxLength: 3000
          description: Full offer description. Optional.
          example: "A luxury hotel in the heart of Paris, steps from the Eiffel Tower."
        comment:
          type: string
          maxLength: 1000
          description: Short owner comment or house rules. Optional.
          example: "Breakfast included. Check-in after 15:00."
        price:
          type: number
          format: double
          minimum: 0
          description: >
            Nightly price in the owner's currency (stored as `BigDecimal`). Must be >= 0.
          example: 249.99
        ownerEmail:
          type: string
          format: email
          description: >
            Email of the offer owner. Set server-side from the JWT on write; present in all
            read responses. Ignored if supplied in a create request body.
          example: "owner@example.com"
        roomCapacity:
          type: integer
          format: int64
          minimum: 1
          description: Maximum number of guests. Must be >= 1.
          example: 2
        city:
          type: string
          description: City where the accommodation is located. Must not be blank.
          example: "Paris"
        country:
          type: string
          description: Country where the accommodation is located. Must not be blank.
          example: "France"
        photoPath:
          type: string
          description: Relative path or URL to the primary photo. Optional.
          example: "/images/grand-hotel-paris.jpg"

    OfferEditDTO:
      type: object
      description: >
        Partial-update payload for `PUT /owner/offers`. Only `id` is required; all other
        fields are optional and merged with the existing stored values via
        `Objects.requireNonNullElseGet`. The `ownerEmail` field is set server-side and
        ignored in the request body.
      required:
        - id
      properties:
        id:
          type: integer
          format: int64
          description: ID of the offer to update. Must exist and be owned by the caller.
          example: 42
        hotelName:
          type: string
          example: "Grand Hotel Paris (Renovated)"
        description:
          type: string
          maxLength: 3000
          example: "Freshly renovated rooms with panoramic city views."
        comment:
          type: string
          maxLength: 1000
          example: "Breakfast and spa access included."
        price:
          type: number
          format: double
          minimum: 0
          example: 299.99
        ownerEmail:
          type: string
          format: email
          description: Ignored in request body; set from JWT by the service.
          example: "owner@example.com"
        roomCapacity:
          type: integer
          format: int64
          minimum: 1
          example: 2
        city:
          type: string
          example: "Paris"
        country:
          type: string
          example: "France"
        photoPath:
          type: string
          example: "/images/grand-hotel-paris-new.jpg"

    PageOfOfferDTO:
      type: object
      description: Spring Data `Page<OfferDTO>` envelope returned by list endpoints.
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/OfferDTO"
        totalElements:
          type: integer
          format: int64
          example: 47
        totalPages:
          type: integer
          example: 3
        size:
          type: integer
          example: 20
        number:
          type: integer
          description: Zero-based current page index.
          example: 0
        first:
          type: boolean
          example: true
        last:
          type: boolean
          example: false
        numberOfElements:
          type: integer
          example: 20
        empty:
          type: boolean
          example: false

    ProblemDetail:
      type: object
      description: >
        RFC 7807 / Spring 6 `ErrorResponse` problem detail. Returned by `GlobalExceptionHandler`
        for validation failures and domain `OfferException` — both mapped to 400.
      properties:
        type:
          type: string
          format: uri
          example: "about:blank"
        title:
          type: string
          example: "Bad Request"
        status:
          type: integer
          example: 400
        detail:
          type: string
          description: >
            Human-readable detail. For validation errors, a semicolon-separated list in the
            form `Field '<name>' <constraint message>`.
          example: "Field 'hotelName' must not be blank; Field 'city' must not be blank"
        instance:
          type: string
          format: uri
          example: "/api/v1/owner/offers"

paths:
  /offers:
    get:
      operationId: getAllOffers
      summary: Get all offers (public, paginated)
      description: >
        Returns all available offers as a paginated list. No authentication required.
        Use `?page=` and `?size=` to control pagination; defaults to page 0, size 20.
      security: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Zero-based page index.
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Number of items per page.
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: Sort expression, e.g. `price,asc`. Repeatable.
      responses:
        "200":
          description: Successful — returns a page of offers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageOfOfferDTO"
              example:
                content:
                  - id: 42
                    hotelName: "Grand Hotel Paris"
                    description: "A luxury hotel in the heart of Paris."
                    comment: "Breakfast included."
                    price: 249.99
                    ownerEmail: "owner@example.com"
                    roomCapacity: 2
                    city: "Paris"
                    country: "France"
                    photoPath: "/images/grand-hotel-paris.jpg"
                totalElements: 47
                totalPages: 3
                size: 20
                number: 0
                first: true
                last: false
                numberOfElements: 20
                empty: false

  /search:
    post:
      operationId: searchOffers
      summary: Search offers by keyword (public)
      description: >
        Full-text keyword search across offers. The request body is a plain string — the
        search term. No authentication required. Returns a flat list (not paginated) of
        matching offers.
      security: []
      requestBody:
        required: true
        content:
          text/plain:
            schema:
              type: string
            example: "Paris"
      responses:
        "200":
          description: Successful — returns matching offers (may be empty).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OfferDTO"
              example:
                - id: 42
                  hotelName: "Grand Hotel Paris"
                  description: "A luxury hotel in the heart of Paris."
                  price: 249.99
                  ownerEmail: "owner@example.com"
                  roomCapacity: 2
                  city: "Paris"
                  country: "France"
                  photoPath: "/images/grand-hotel-paris.jpg"

  /owner/offers:
    get:
      operationId: getOwnedOffers
      summary: Get offers owned by the authenticated user (paginated)
      description: >
        Returns all offers created by the user identified by the JWT `email` claim.
        Use `?page=` and `?size=` to control pagination; defaults to page 0, size 20.
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
          description: Zero-based page index.
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Number of items per page.
        - name: sort
          in: query
          required: false
          schema:
            type: string
          description: Sort expression, e.g. `price,asc`. Repeatable.
      responses:
        "200":
          description: Successful — returns a page of the owner's offers.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageOfOfferDTO"
              example:
                content:
                  - id: 42
                    hotelName: "Grand Hotel Paris"
                    price: 249.99
                    ownerEmail: "owner@example.com"
                    roomCapacity: 2
                    city: "Paris"
                    country: "France"
                totalElements: 1
                totalPages: 1
                size: 20
                number: 0
                first: true
                last: true
                numberOfElements: 1
                empty: false
        "401":
          description: Missing or invalid bearer token.

    post:
      operationId: createOffer
      summary: Create a new offer
      description: >
        Creates an offer owned by the authenticated user. The `ownerEmail` field in the
        request body is ignored; the service sets it from the JWT `email` claim. Returns
        201 on success and publishes a Kafka notification event.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferDTO"
            example:
              hotelName: "Grand Hotel Paris"
              description: "A luxury hotel in the heart of Paris, steps from the Eiffel Tower."
              comment: "Breakfast included."
              price: 249.99
              roomCapacity: 2
              city: "Paris"
              country: "France"
              photoPath: "/images/grand-hotel-paris.jpg"
      responses:
        "201":
          description: Offer created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDTO"
              example:
                id: 42
                hotelName: "Grand Hotel Paris"
                description: "A luxury hotel in the heart of Paris, steps from the Eiffel Tower."
                comment: "Breakfast included."
                price: 249.99
                ownerEmail: "owner@example.com"
                roomCapacity: 2
                city: "Paris"
                country: "France"
                photoPath: "/images/grand-hotel-paris.jpg"
        "400":
          description: >
            Validation failure (`hotelName`, `price`, `roomCapacity`, `city`, or `country`
            missing/invalid) or domain-level `OfferException`.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
              example:
                type: "about:blank"
                title: "Bad Request"
                status: 400
                detail: "Field 'hotelName' must not be blank"
                instance: "/api/v1/owner/offers"
        "401":
          description: Missing or invalid bearer token.

    put:
      operationId: editOffer
      summary: Edit an existing offer
      description: >
        Partially updates an offer. Only fields present and non-null in the request body
        are applied; missing fields retain their stored values. The `id` field is required.
        The `ownerEmail` field is set from the JWT `email` claim regardless of what the
        caller sends. Returns 200 with the full updated offer.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferEditDTO"
            example:
              id: 42
              hotelName: "Grand Hotel Paris (Renovated)"
              price: 299.99
      responses:
        "200":
          description: Offer updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OfferDTO"
              example:
                id: 42
                hotelName: "Grand Hotel Paris (Renovated)"
                description: "A luxury hotel in the heart of Paris, steps from the Eiffel Tower."
                price: 299.99
                ownerEmail: "owner@example.com"
                roomCapacity: 2
                city: "Paris"
                country: "France"
        "400":
          description: >
            Validation failure (`id` missing, `price` < 0, `roomCapacity` < 1) or domain-level
            `OfferException` (e.g. offer not found, not owned by caller).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid bearer token.

  /owner/offers/{offerId}:
    delete:
      operationId: deleteOffer
      summary: Delete an offer
      description: >
        Deletes the offer with the given ID. The service verifies ownership via the JWT
        `email` claim. Returns 200 with a JSON-encoded confirmation string on success.
        The `offerId` path parameter is parsed as `Long` internally.
      security:
        - bearerAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: string
          description: ID of the offer to delete (numeric, passed as string).
          example: "42"
      responses:
        "200":
          description: Offer deleted. Body is a JSON-encoded string.
          content:
            application/json:
              schema:
                type: string
              example: "\"Offer with id: 42 deleted.\""
        "400":
          description: Domain-level `OfferException` (e.g. offer not found, not owned by caller).
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "401":
          description: Missing or invalid bearer token.

  /offers/{offerId}/owner:
    get:
      operationId: getOfferOwner
      summary: Get the owner email for an offer
      description: >
        Returns the owner email for the given offer ID. Used by sky-booking as a
        service-to-service call when creating a booking. Requires a valid bearer JWT.
        The `offerId` path parameter is parsed as `Long` internally.
      security:
        - bearerAuth: []
      parameters:
        - name: offerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Numeric ID of the offer whose owner to look up.
          example: 42
      responses:
        "200":
          description: Owner email returned as a plain string.
          content:
            application/json:
              schema:
                type: string
              example: "owner@example.com"
        "400":
          description: Non-numeric offerId or other bad request.
        "401":
          description: Missing or invalid bearer token.